extract types and related util to a new module
