+++ static/js/browser2/statvar_hierarchy_search.tsx
@@ -22,7 +22,7 @@
 import React from "react";
 import axios from "axios";
 import _ from "lodash";
-import { StatVarGroupNodeType, StatVarNodeType } from "./util";
+import { StatVarGroupNodeType, StatVarNodeType } from "./types";
 
 interface StatVarHierarchySearchPropType {
   statVarGroupsData: { [key: string]: StatVarGroupNodeType };
