+++ static/js/browser2/weather_chart_section.tsx
@@ -22,7 +22,8 @@ import React from "react";
 import axios from "axios";
 import _ from "lodash";
 import { ObservationChart } from "./observation_chart";
-import { getUnit, loadSpinner, removeSpinner, SourceSeries } from "./util";
+import { getUnit, loadSpinner, removeSpinner } from "./util";
+import { SourceSeries } from "./types";
 
 const WEATHER_PROPERTY_NAMES = [
   "temperature",
